refactor selection model
