+++ docs/sportsbooklib/calculators/parlay_calc.md
@@ -5,7 +5,7 @@ Functions
 ---------
 
     
-`get_parlay_payout(stake: decimal.Decimal, selections: List[sportsbooklib.models.selection.selection.Selection]) ‑> List[Mapping[str, Union[decimal.Decimal, sportsbooklib.models.odds.Odds]]]`
+`get_parlay_payout(stake: decimal.Decimal, selections: List[sportsbooklib.models.selection.Selection]) ‑> List[Mapping[str, Union[decimal.Decimal, sportsbooklib.models.odds.Odds]]]`
 :   Given stake and list of n odds, calculate combined odds and return
     
     Parameters
