+++ docs/sportsbooklib/models/selection/index.md
@@ -2,6 +2,15 @@ Module sportsbooklib.models.selection
 =====================================
 The selection submodule.
 
-Sub-modules
------------
-* sportsbooklib.models.selection.selection
+Classes
+-------
+
+`Selection(name: str, odds: sportsbooklib.models.odds.Odds)`
+:   The Selection Object.
+    
+    Attributes
+    ----------
+    name : str
+        Name of the selection.
+    odds : Odds
+        The Odds object associated with the selection.
