+++ sportsbooklib/calculators/parlay_calc.py
@@ -5,7 +5,7 @@ from sportsbooklib.models.exceptions import InvalidSelectionsInputException
 from sportsbooklib.models.odds.enums import OddsFormat
 from sportsbooklib.models.odds import Odds
 
-from sportsbooklib.models.selection.selection import Selection
+from sportsbooklib.models.selection import Selection
 
 
 def get_parlay_payout(stake: Decimal, selections: List[Selection]) -> List[Mapping[str, Union[Decimal, Odds]]]:
