+++ sportsbooklib/models/selection/__init__.py
@@ -1,3 +1,23 @@
 """
 The selection submodule.
 """
+
+from sportsbooklib.models.odds import Odds
+
+
+class Selection:
+    """
+
+    The Selection Object.
+
+    Attributes
+    ----------
+    name : str
+        Name of the selection.
+    odds : Odds
+        The Odds object associated with the selection.
+    """
+
+    def __init__(self, name: str, odds: Odds):
+        self.name = name
+        self.odds = odds
