+++ tests/calculators/test_parlay_calc.py
@@ -6,7 +6,7 @@ from decimal import Decimal
 from sportsbooklib.models.exceptions import InvalidSelectionsInputException
 from sportsbooklib.models.odds.enums import OddsFormat
 from sportsbooklib.models.odds import Odds
-from sportsbooklib.models.selection.selection import Selection
+from sportsbooklib.models.selection import Selection
 
 
 def test_invalid_empty_selections():
