+++ tests/models/selection/test_selection.py
@@ -1,4 +1,4 @@
-from sportsbooklib.models.selection.selection import Selection
+from sportsbooklib.models.selection import Selection
 from sportsbooklib.models.odds import Odds
 from sportsbooklib.models.odds.enums import OddsFormat
